Branch ResponsivePicture away from ResponsiveImage
